fix: topbar, logo, date chips, Bagre title, conquistas display
- Remove 54px status bar spacer from topbar; use env(safe-area-inset-top)
- Increase logo from 32→48px
- Remove overflow:hidden causing date/time chip clipping
- Fix Bagre da Noite title wrapping (remove whiteSpace:nowrap, add left/right constraints)
- Shorten date format to weekday-short + day + month-short
- Replace broken medal PNG fallback with trait emoji display (traitNome ≠ medal names in DB)

+++ src/app/feed/page.tsx
@@ -86,7 +86,7 @@ export default async function FeedPage() {
     : false;
 
   const dataRodada = rodadaAtiva
-    ? new Date(rodadaAtiva.data).toLocaleDateString("pt-BR", { weekday: "long", day: "numeric", month: "long" })
+    ? new Date(rodadaAtiva.data).toLocaleDateString("pt-BR", { weekday: "short", day: "numeric", month: "short" })
     : null;
 
   // Last 3 rodada dates for filter pills
